fix stores
